testing/web-platform/tests/XMLHttpRequest/withcredentials-set.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: setting withCredentials and reading value back</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-withcredentials-attribute" data-tested-assertations="following::ol/li[5]" />  
  </head>
  <body>
    <div id="log"></div>
    <script>
      var client = new XMLHttpRequest()
      test(function() {
        assert_true(client.withCredentials === false, "withCredentials defaults to false")
        client.withCredentials=true
        assert_true(client.withCredentials, "is true after setting")
      }, "default value is false, set value is true")

      var client = new XMLHttpRequest()
      test(function() {
        client.open("GET", "resources/delay.py?ms=1000", false)
        client.withCredentials=true
        assert_true(client.withCredentials, "set in OPEN state")
      }, "can also be set in OPEN state")

    </script>
  </body>
</html>
